requirements.txt
numpy
pandas
pyyaml
akshare
baostock
pytdx
scipy
scikit-learn
arch
